+++ shuup_tests/core/test_orders.py
@@ -495,21 +495,22 @@ def test_refund_with_shipment(restock):
 def test_refund_entire_order_restock_shipment_no_supplier_module():
     shop = get_default_shop()
     supplier = get_default_supplier()
+    supplier.supplier_modules.clear()
     product = create_product(
         "test-sku",
         shop=get_default_shop(),
         default_price=10,
     )
-    check_stock_counts(supplier, product, 0, 0)
+    assert not supplier.get_stock_statuses([product.id])
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
     product_line = order.lines.first()
     order.create_shipment({product_line.product: 2}, supplier=supplier)
-    check_stock_counts(supplier, product, 0, 0)
+    assert not supplier.get_stock_statuses([product.id])
 
     # Create a full refund with `restock_products` set to True
     order.create_full_refund(restock_products=True)
 
-    check_stock_counts(supplier, product, 0, 0)
+    assert not supplier.get_stock_statuses([product.id])
 
 
 @pytest.mark.django_db