+++ CHANGELOG.md
@@ -11,6 +11,7 @@ List all changes after the last release here (newer on top). Each change on a se
 ### Changed
 
 - Admin: filter products in list and edit views according to product kind listing name
+- Core: suppliers don't have default supplier modules anymore
 
 ### Add